+++ lib/ebook_renamer/errors.rb
@@ -1,0 +1,5 @@
+module EbookRenamer
+  module Errors
+    EbookMetaNotInstall = Class.new(StandardError)
+  end
+end